barely
