+++ frontend/src/pages/Courses/year/semester/add/AddButton.module.css
@@ -62,33 +62,36 @@
 /* */
 
 .TermOptions {
-	position: absolute;
-	background-color: #fff;
-	border: 1px solid #ccc;
-	border-radius: 4px;
-	top: 100%;
-	left: 0;
-	margin-top: 5px;
-	z-index: 9999;
-	width: 90px;
-	box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-	max-height: 200px;
-	overflow-y: auto;
-	font-size: 12px;
-	box-sizing: border-box;
+  position: absolute;
+  background-color: #fff;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  top: 100%;
+  left: 0;
+  margin-top: 5px;
+  z-index: 9999;
+  width: 90px;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+  max-height: 200px;
+  overflow-y: auto;
+  font-size: 12px;
+  box-sizing: border-box;
 }
+
 .TermOptions div {
-	padding: 8px;
-	cursor: pointer;
-	color: black;
-	background-color: white;
+  padding: 8px;
+  cursor: pointer;
+  color: black;
+  background-color: white;
 }
+
 .TermOptions div:hover {
-	background-color: #f0f0f0;
+  background-color: #f0f0f0;
 }
 
 .SelectedTerm {
-  background-color: #f0f0f0;
+  background-color: #93c7ff !important; /* Light blue background for the selected term */
+  color: black; /* Optional: change text color for selected term */
 }
 
 /* */
